@86erh246f #Moved the google recaptcha url to config
